Moving the Saying representation to the api package
